Fixed Khiva path and Local maximals tests

+++ KhivaCsharp/KhivaCSharpNUnitTest/FeaturesTests.cs
@@ -541,21 +541,19 @@ namespace khiva.features.tests
                 }
             }
         }
-        /*
-         * TODO
-         * Descomentar este bloque de código cuando se arregle la función en khiva
+       
         [Test]
         public void TestLocalMaximals()
         {
             float[,] tss = { { 0.0F, 4.0F, 3.0F, 5.0F, 4.0F, 1.0F, 0.0F, 4.0F }, { 0.0F, 4.0F, 3.0F, 5.0F, 4.0F, 1.0F, 0.0F, 4.0F } };
             using(KhivaArray arr = new KhivaArray(tss), localMaximals = Features.LocalMaximals(arr))
             {
-                int[,] result = localMaximals.GetData2D<int>();
+                float[,] result = localMaximals.GetData2D<float>();
                 float[,] expected = { { 0.0F, 1.0F, 0.0F, 1.0F, 0.0F, 0.0F, 0.0F, 1.0F }, { 0.0F, 1.0F, 0.0F, 1.0F, 0.0F, 0.0F, 0.0F, 1.0F } };
                 Assert.AreEqual(expected, result);
             }
         }
-        */
+        
         [Test]
         public void TestLongestStrikeAboveMean()
         {
